requirements.txt
asgiref==3.5.2
black==22.3.0
click==8.1.3
colorama==0.4.4
dj-database-url==0.5.0
Django==4.0.5
django-ckeditor==6.4.2
django-heroku==0.3.1
django-js-asset==2.0.0
gunicorn==20.1.0
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==9.1.1
platformdirs==2.5.2
psycopg==3.0.14
psycopg2==2.9.3
sqlparse==0.4.2
tomli==2.0.1
tzdata==2022.1
whitenoise==6.2.0
